+++ CHANGES.rst
@@ -6,6 +6,7 @@ Dev
 ===
 
 - Add ``compilestatic`` management command (replaces ``static_precompiler_watch``)
+- Add ``STATIC_PRECOMPILER_DISABLE_AUTO_COMPILE`` to settings
 - Add ``STATIC_PRECOMPILER_CACHE_NAME`` to settings
 
 0.6